+++ src/sc_keeper/references.py
@@ -23,6 +23,9 @@ from sqlmodel import select
 
 from .database import session
 
+# make sure we have a fresh database
+session.updated.wait()
+
 # create enums from DB values for filtering options
 with session.sessionmaker as db:
     Countries = StrEnum(
